MEP Matrix: Add missing functions
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@22746 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 82f52f784a43af81c3ef4ff3d563770001d487a4

+++ common/mep.matrix/src/de/uniol/inf/is/odysseus/mep/matrix/CountVectorFunction.java
@@ -32,7 +32,7 @@ public class CountVectorFunction extends AbstractFunction<Integer> {
     private static final SDFDatatype[][] ACC_TYPES = new SDFDatatype[][] { SDFDatatype.VECTORS };
 
     public CountVectorFunction() {
-        super("sCount", 1, CountVectorFunction.ACC_TYPES, SDFDatatype.DOUBLE);
+        super("sCount", 1, CountVectorFunction.ACC_TYPES, SDFDatatype.INTEGER);
     }
 
     @Override
